Update appView.js
If you do `this.listenTo(this.collection, "all", this.render());` then you're setting the **result** of `this.render` as callback when an event fires on the collection. As `this.render` returns `this`, that means that Backbone will try to use the view object itself as a callback, which will blow up. The fix is to use `this.render` as callback instead of `this.render()`.